event_mgmt/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-flex-fields==0.9.7
importlib-metadata==4.11.2
Markdown==3.3.6
Pillow==9.0.1
psycopg2==2.9.3
PyJWT==2.3.0
python-magic==0.4.25
pytz==2021.3
sqlparse==0.4.2
zipp==3.7.0
